backend/requirements.txt
uvicorn
fastapi[all]
python-jose
passlib
pymongo
apscheduler
resend
pytz